requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pydantic-settings
sqlalchemy
aiosqlite
python-dotenv
solders
solana
python-jose[cryptography]
passlib[bcrypt]
python-multipart
shortuuid
qrcode[pil]
Pillow
